New payment process (part 1)
